refactor: 统一认证机制为依赖注入，移除 BaseHTTPMiddleware
