+++ tests/test_auth_middleware.py
@@ -1,36 +1,55 @@
+import os
 import unittest
+from unittest.mock import MagicMock, patch
 
-import tests._pathsetup  # noqa: F401  (adds backend/app to sys.path)
+import tests._pathsetup  # noqa: F401
 
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
+from starlette.status import HTTP_401_UNAUTHORIZED, HTTP_503_SERVICE_UNAVAILABLE
 
-from utils.security import ApiKeyMiddleware
+from utils.security import require_configured_api_key
+from router.manage import manage_app
 
 
-class ApiKeyMiddlewareTest(unittest.TestCase):
-    def _build_app(self, api_key):
-        app = FastAPI()
+class RequireConfiguredApiKeyTest(unittest.TestCase):
+    def test_returns_503_when_api_key_not_configured(self):
+        with patch.dict(os.environ, {'CUITCCA_API_KEY': ''}):
+            app = FastAPI()
+            app.include_router(manage_app, prefix='/manage')
+            client = TestClient(app)
+            response = client.get("/manage/stats")
+            self.assertEqual(response.status_code, HTTP_503_SERVICE_UNAVAILABLE)
 
-        @app.get("/")
-        def root():
-            return {"ok": True}
+    def test_returns_401_when_token_missing(self):
+        with patch.dict(os.environ, {'CUITCCA_API_KEY': 'secret123'}):
+            app = FastAPI()
+            app.include_router(manage_app, prefix='/manage')
+            client = TestClient(app)
+            response = client.get("/manage/stats")
+            self.assertEqual(response.status_code, HTTP_401_UNAUTHORIZED)
 
-        app.add_middleware(ApiKeyMiddleware, api_key=api_key)
-        return app
+    def test_returns_401_when_token_wrong(self):
+        with patch.dict(os.environ, {'CUITCCA_API_KEY': 'secret123'}):
+            app = FastAPI()
+            app.include_router(manage_app, prefix='/manage')
+            client = TestClient(app)
+            response = client.get(
+                "/manage/stats",
+                headers={"Authorization": "Bearer wrong"}
+            )
+            self.assertEqual(response.status_code, HTTP_401_UNAUTHORIZED)
 
-    def test_rejects_missing_token_with_a_clean_401_not_a_500(self):
-        """Raising HTTPException inside BaseHTTPMiddleware.dispatch() isn't caught
-        by FastAPI's exception handlers and surfaces as a raw 500 instead of the
-        intended 401 -- verified empirically against a minimal repro app."""
-        client = TestClient(self._build_app("secret123"), raise_server_exceptions=False)
-        response = client.get("/")
-        self.assertEqual(response.status_code, 401)
-
-    def test_accepts_correct_token(self):
-        client = TestClient(self._build_app("secret123"), raise_server_exceptions=False)
-        response = client.get("/", headers={"Authorization": "Bearer secret123"})
-        self.assertEqual(response.status_code, 200)
+    def test_accepts_correct_bearer_token(self):
+        with patch.dict(os.environ, {'CUITCCA_API_KEY': 'secret123'}):
+            app = FastAPI()
+            app.include_router(manage_app, prefix='/manage')
+            client = TestClient(app)
+            response = client.get(
+                "/manage/stats",
+                headers={"Authorization": "Bearer secret123"}
+            )
+            self.assertEqual(response.status_code, 200)
 
 
 if __name__ == '__main__':
